+++ .github/workflows/build-bindings-android.yml
@@ -21,14 +21,35 @@ on:
         required: false
         type: boolean
         default: false
+      uniffi-25:
+        description: 'If true, builds additional bindings for Uniffi 0.25'
+        required: false
+        type: boolean
+        default: false
 
 jobs:
+  setup:
+    name: Setup
+    runs-on: ubuntu-latest
+    steps:
+      - id: set-matrix
+        run: |
+          if [ ${{ inputs.uniffi-25 }} == true ]; then
+            echo "::set-output name=matrix::['', '-uniffi-25']"
+          else
+            echo "::set-output name=matrix::['']"
+          fi
+    outputs:
+      uniffi-matrix: ${{ steps.set-matrix.outputs.matrix }}
+
   build:
     if: ${{ !inputs.use-dummy-binaries }}
     runs-on: ubuntu-latest
-    name: build ${{ matrix.target }}
+    name: build ${{ matrix.target }}${{ matrix.uniffi }}
+    needs: setup
     strategy:
       matrix:
+        uniffi: ${{ fromJson(needs.setup.outputs.uniffi-matrix) }}
         target: [
           aarch64-linux-android,
           armv7-linux-androideabi,
@@ -66,12 +87,18 @@ jobs:
 
     - uses: Swatinem/rust-cache@v2
       with:
+        key: ${{ matrix.uniffi }}
         workspaces: libs
 
     - name: Build sdk-bindings
+      if: matrix.uniffi != '-uniffi-25'
       working-directory: libs/sdk-bindings
-      run: |
-        cargo ndk -t ${{ matrix.target }} build --release
+      run: cargo ndk -t ${{ matrix.target }} build --release
+
+    - name: Build sdk-bindings Uniffi 0.25
+      if: matrix.uniffi == '-uniffi-25'
+      working-directory: libs/sdk-bindings
+      run: cargo ndk -t ${{ matrix.target }} build --no-default-features --features=uniffi-25 --release
 
     - name: Copy build output
       run: |
@@ -89,46 +116,53 @@ jobs:
     - name: Archive release
       uses: actions/upload-artifact@v4
       with:
-        name: sdk-bindings-${{ matrix.target }}
+        name: sdk-bindings-${{ matrix.target }}${{ matrix.uniffi }}
         path: dist/*
   
   jnilibs:
-    needs: build
+    needs: 
+    - build
+    - setup
     runs-on: ubuntu-latest
-    name: build jniLibs
+    name: build jniLibs${{ matrix.uniffi }}
+    strategy:
+      matrix:
+        uniffi: ${{ fromJson(needs.setup.outputs.uniffi-matrix) }}
     steps:
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-aarch64-linux-android
+          name: sdk-bindings-aarch64-linux-android${{ matrix.uniffi }}
           path: arm64-v8a
       
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-armv7-linux-androideabi
+          name: sdk-bindings-armv7-linux-androideabi${{ matrix.uniffi }}
           path: armeabi-v7a
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-i686-linux-android
+          name: sdk-bindings-i686-linux-android${{ matrix.uniffi }}
           path: x86
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-x86_64-linux-android
+          name: sdk-bindings-x86_64-linux-android${{ matrix.uniffi }}
           path: x86_64
       
       - name: Archive jniLibs
         uses: actions/upload-artifact@v4
         with:
-          name: sdk-bindings-android-jniLibs
+          name: sdk-bindings-android-jniLibs${{ matrix.uniffi }}
           path: ./*
 
   build-dummies:
     if: ${{ inputs.use-dummy-binaries }}
     runs-on: ubuntu-latest
-    name: build android dummies
+    name: build dummies ${{ matrix.target }}${{ matrix.uniffi }}
+    needs: setup
     strategy:
       matrix:
+        uniffi: ${{ fromJson(needs.setup.outputs.uniffi-matrix) }}
         target: [
           aarch64-linux-android,
           armv7-linux-androideabi,
@@ -144,36 +178,41 @@ jobs:
       - name: Upload dummy Android ${{ matrix.target }} artifact
         uses: actions/upload-artifact@v4
         with:
-          name: sdk-bindings-${{ matrix.target }}
+          name: sdk-bindings-${{ matrix.target }}${{ matrix.uniffi }}
           path: ./*
 
   jnilibs-dummy:
-    needs: build-dummies
+    needs:
+    - setup
+    - build-dummies
     runs-on: ubuntu-latest
-    name: build jniLibs dummy
+    name: build jniLibs dummy ${{ matrix.uniffi }}
+    strategy:
+      matrix:
+        uniffi: ${{ fromJson(needs.setup.outputs.uniffi-matrix) }}
     steps:
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-aarch64-linux-android
+          name: sdk-bindings-aarch64-linux-android${{ matrix.uniffi }}
           path: arm64-v8a
       
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-armv7-linux-androideabi
+          name: sdk-bindings-armv7-linux-androideabi${{ matrix.uniffi }}
           path: armeabi-v7a
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-i686-linux-android
+          name: sdk-bindings-i686-linux-android${{ matrix.uniffi }}
           path: x86
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-x86_64-linux-android
+          name: sdk-bindings-x86_64-linux-android${{ matrix.uniffi }}
           path: x86_64
       
       - name: Archive jniLibs
         uses: actions/upload-artifact@v4
         with:
-          name: sdk-bindings-android-jniLibs
+          name: sdk-bindings-android-jniLibs${{ matrix.uniffi }}
           path: ./*
