+++ .github/workflows/publish-android.yml
@@ -39,7 +39,7 @@ jobs:
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-android-jniLibs
+          name: sdk-bindings-android-jniLibs-uniffi-25
           path: libs/sdk-bindings/bindings-android/lib/src/main/jniLibs
 
       - uses: actions/download-artifact@v4
