+++ .github/workflows/publish-golang.yml
@@ -38,47 +38,47 @@ jobs:
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-i686-linux-android
+          name: sdk-bindings-i686-linux-android-uniffi-25
           path: breez_sdk/lib/android-386
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-armv7-linux-androideabi
+          name: sdk-bindings-armv7-linux-androideabi-uniffi-25
           path: breez_sdk/lib/android-aarch
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-aarch64-linux-android
+          name: sdk-bindings-aarch64-linux-android-uniffi-25
           path: breez_sdk/lib/android-aarch64
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-x86_64-linux-android
+          name: sdk-bindings-x86_64-linux-android-uniffi-25
           path: breez_sdk/lib/android-amd64
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-aarch64-apple-darwin
+          name: sdk-bindings-aarch64-apple-darwin-uniffi-25
           path: breez_sdk/lib/darwin-aarch64
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-x86_64-apple-darwin
+          name: sdk-bindings-x86_64-apple-darwin-uniffi-25
           path: breez_sdk/lib/darwin-amd64
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-aarch64-unknown-linux-gnu
+          name: sdk-bindings-aarch64-unknown-linux-gnu-uniffi-25
           path: breez_sdk/lib/linux-aarch64
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-x86_64-unknown-linux-gnu
+          name: sdk-bindings-x86_64-unknown-linux-gnu-uniffi-25
           path: breez_sdk/lib/linux-amd64
 
       - uses: actions/download-artifact@v4
         with:
-          name: sdk-bindings-x86_64-pc-windows-msvc
+          name: sdk-bindings-x86_64-pc-windows-msvc-uniffi-25
           path: breez_sdk/lib/windows-amd64
 
       - name: Archive Go release
@@ -94,6 +94,8 @@ jobs:
         run: |
           git config --global user.email github-actions@github.com
           git config --global user.name github-actions
+          git add breez_sdkbreez_sdk.h
+          git add breez_sdk/breez_sdk.c
           git add breez_sdk/breez_sdk.go
           git add breez_sdk/lib/android-386/libbreez_sdk_bindings.so
           git add breez_sdk/lib/android-386/libc++_shared.so
