+++ libs/sdk-bindings/bindings-kotlin-multiplatform/breez-sdk-kmp/src/nativeInterop/cinterop/breez.def
@@ -3,10 +3,10 @@ package = breez_sdk.cinterop
 
 staticLibraries = libbreez_sdk_bindings.a
 
-libraryPaths.ios_arm64 = src/libs/ios-arm64
-libraryPaths.ios_simulator_arm64 = src/libs/ios-simulator-arm64
-libraryPaths.ios_simulator_x64 = src/libs/ios-simulator-x64
-libraryPaths.ios_x64 = src/libs/ios-simulator-x64
+libraryPaths.ios_arm64 = src/lib/ios-arm64
+libraryPaths.ios_simulator_arm64 = src/lib/ios-simulator-arm64
+libraryPaths.ios_simulator_x64 = src/lib/ios-simulator-x64
+libraryPaths.ios_x64 = src/lib/ios-simulator-x64
 
 
 linkerOpts = -ios_simulator_version_min 14.0.0 -framework Security
