+++ libs/sdk-bindings/tests/test_generated_bindings.rs
@@ -1,5 +1,10 @@
+#[cfg(feature = "uniffi-28")]
+extern crate uniffi_28 as uniffi;
+
+#[cfg(feature = "uniffi-25")]
 use std::process::Command;
 
+#[cfg(feature = "uniffi-28")]
 uniffi::build_foreign_language_testcases!(
     "tests/bindings/test_breez_sdk.swift",
     "tests/bindings/test_breez_sdk.kts",
@@ -7,6 +12,7 @@ uniffi::build_foreign_language_testcases!(
 );
 
 #[test]
+#[cfg(feature = "uniffi-25")]
 fn test_csharp() {
     let output = Command::new("dotnet")
         .arg("run")
@@ -21,6 +27,7 @@ fn test_csharp() {
 }
 
 #[test]
+#[cfg(feature = "uniffi-25")]
 fn test_golang() {
     let output = Command::new("go")
         .env(
