+++ libs/sdk-core/Cargo.toml
@@ -7,6 +7,12 @@ version.workspace = true
 name = "breez_sdk_core"
 crate-type = ["staticlib", "cdylib", "lib"]
 
+[features]
+default = []
+# Uniffi features required to build using cargo-lipo
+uniffi-25 = []
+uniffi-28 = []
+
 [dependencies]
 flutter_rust_bridge = "=1.82.6"
 aes = { workspace = true }
